+++ tap_freshdesk/discover.py
@@ -3,10 +3,37 @@ from singer import metadata
 from singer.catalog import Catalog, CatalogEntry, Schema
 from tap_freshdesk.schema import get_schemas
 from tap_freshdesk.streams import STREAMS
-from tap_freshdesk.exceptions import freshdeskUnauthorizedError, freshdeskForbiddenError
-from tap_freshdesk.utils import check_stream_access
+from tap_freshdesk.exceptions import freshdeskUnauthorizedError, freshdeskForbiddenError, freshdeskNoAccessibleStreamsError
 
 LOGGER = singer.get_logger()
+
+
+def check_stream_access(stream_name, probe_fn, auth_error_types, fallback_accessible=False):
+    """
+    Probe a stream endpoint and return True if accessible, False on auth error.
+
+    :param stream_name: Used in log messages.
+    :param probe_fn: Zero-argument callable that performs the API probe.
+    :param auth_error_types: Exception type(s) indicating 401/403 — returns False.
+    :param fallback_accessible: If True, non-auth errors (e.g. 400 from minimal
+                                probe params) are treated as auth-OK and return True.
+                                If False (default), they are re-raised.
+    """
+    try:
+        probe_fn()
+        LOGGER.info("Stream '%s' is accessible.", stream_name)
+        return True
+    except auth_error_types:
+        LOGGER.warning(
+            "Stream '%s' is not accessible with the provided credentials.",
+            stream_name,
+        )
+        return False
+    except Exception:  # pylint: disable=broad-except
+        if fallback_accessible:
+            LOGGER.info("Stream '%s' endpoint reachable (auth OK).", stream_name)
+            return True
+        raise
 
 
 def _check_stream_access(client, stream_name, stream_class) -> bool:
@@ -42,6 +69,10 @@ def discover(client) -> Catalog:
 
     for stream_name, schema_dict in schemas.items():
         if not _check_stream_access(client, stream_name, STREAMS[stream_name]):
+            LOGGER.warning(
+                "Stream '%s' will be excluded from the catalog due to insufficient permissions.",
+                stream_name,
+            )
             continue
 
         try:
@@ -63,6 +94,12 @@ def discover(client) -> Catalog:
                 schema=schema,
                 metadata=mdata,
             )
+        )
+
+    if not catalog.streams:
+        raise freshdeskNoAccessibleStreamsError(
+            "No stream endpoints are accessible with the provided credentials. "
+            "Verify that the API key has the required permissions."
         )
 
     return catalog
